# 8-week/api/gen/openapi/gateway.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Auth API",
    "description": "RESTful gateway API for authentication and tokens.",
    "version": "1.0.0",
    "contact": {
      "name": "Daniil Kalts",
      "email": "gopher@gmail.com"
    }
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "summary": "Get current profile",
        "description": "Retrieves profile details for the authenticated user.",
        "operationId": "getMyProfile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetProfileResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "🙋 Profile"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete current profile",
        "description": "Deletes the authenticated user's account profile.",
        "operationId": "deleteMyProfile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "🙋 Profile"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "summary": "Update current profile",
        "description": "Partially updates name, email, or password for the authenticated user profile.",
        "operationId": "updateMyProfile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpdateProfileRequest"
            }
          }
        ],
        "tags": [
          "🙋 Profile"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "userV1User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/v1Role"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "v1GetProfileResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/userV1User"
        }
      }
    },
...
    "v1Role": {
      "type": "string",
      "enum": [
        "USER",
        "ADMIN"
      ],
      "default": "USER"
    },
...
    "v1UpdateProfileRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "description": "JWT access token. Example: Bearer \u003ctoken\u003e",
      "name": "Authorization",
      "in": "header"
    }
  }
}